requirements/apps.txt
Django==1.4
django-tastypie==0.9.11
decorator==3.3.2
PIL
-e git+http://github.com/pennersr/django-allauth.git#egg=django-allauth
-e git+http://github.com/pythonforfacebook/facebook-sdk.git#egg=facebook-python-sdk
-e git+http://github.com/ericflo/django-avatar.git#egg=django-avatar

